Add files via upload
This program reads in the contact information in the given data file (contacts.txt). Dynamic allocation is used to allocate data structures for the contact information as the data is read in. Strtok() is used to parse the information in the data file. The data is then printed out using a function which is given the pointer to the first structure in the linked list.